rename common to internal
