+++ plugins/golang-filter/mcp-server/internal/server.go
@@ -1,4 +1,4 @@
-package common
+package internal
 
 import (
 	"context"
